Upgrade to AlfredSwift version with better version parsing
AlfredGifBrowser.app stopped working with Alfred v4.5. The bug was
in the AlfredSwift library.  Before `v0.1.4`, AlfredSwift expected
both minor and patch versions to be present in order to parse
a version-string correctly. Thus, AlfredSwift considered `4.5`
to be an invalid version. This, in turn made AlfredGifBrowser crash.
From `v0.1.4` onwards, AlfredSwift correctly parses `4.5` as
`4.5.0`. Thus, upgrading to it fixes #8 and makes the GIF browser
work again.